commit generated data files
Even though the files are generated by Rake tasks, I think it's a good
idea to commit them, because their generation depends on files from an
external codebase. This way we are more likely to notice destructive
structure changes in the upstream codebase and we always have a backup
of the last valid output.